헤더, 메인페이지 슬라이더, 마이페이지_갤러리 수정
