+++ src/views/gallery/galleryDetail.vue
@@ -3,18 +3,16 @@
         <Header :parent_id="parent_id"/>
         <Gallery_Detail :parent_id="parent_id" />
         <Comment :parent_id="parent_id" />
-        <Footer />
     </div>
 </template>
 
 <script>
 import Header from '@/components/main/header_nav.vue';
-import Footer from '@/components/main/footer_info.vue';
 import Gallery_Detail from '@/components/gallery/gallery_detail.vue';
 import Comment from '@/components/gallery/gallery_comment.vue';
 export default {
     components: {
-        Header, Footer, Gallery_Detail, Comment
+        Header, Gallery_Detail, Comment
     },
     props: {
         parent_id: {
